apps/py/requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.24.0
python-jose==3.3.0
reportlab==4.0.8
requests==2.31.0
